move everything into release.mts, rename release-utils.spec
